Add de-dup for a corner case of compress=true & skipDedup=true
Handle corner case where compressed schemas contain an enum used in two
places but have the same name. This is a case where the enum name should
be deduped, as otherwise an underscore will be added for one of them,
subtracting from the original goal of improving usability.

This change required processing the skipEnumDedup flag when processing
unions as well, as this change makes it so that resolveEnumName requires
skipEnumDedup is always the same between different calls to it.

More tests are added to test the combination of compress + skipDedup
options.