Split genState into smaller generators.
Currently, genState contains all the internal state variables and
generation functions for both Go and proto, making the code more
confusing than necessary and changes to the code seemingly more
precarious due to the lack of clear separation between Go and proto
generation. This is a problem for both the reader of the code as well as
maintainers.

To help with this, genState is split into protoGenState, goGenState,
enumGenState, and genHelper. genHelper contains helpers used for
generation, and enumGenState contains functionality and internal state
used for generating enum names. These two are language independent.
protoGenState and goGenState then contains the language-specific
generation functionalities.

The generation methods are partitioned either into one of these new
smaller generators, or became independent of generation state and lost
their receivers, making each of the generators easier to understand.