Add SplitFiles method to GeneratedPathCode (#342)
Add SplitFiles method to GeneratedPathCode

In order to allow consumers to roughly limit the size of each file.

- TestGeneratePathCodeSplitFiles tests using the openconfig-simple schema, which contains **4 directory entries**: 
  - Device
  - Parent
  - Parent_Child
  - RemoteContainer

These 4 are split between the various files.

I have tested on the real OpenConfig schema and splitting by counting structs does have a large variance, but it's a good enough approximation. Counting bytes is more accurate, but I think it's slightly more complex, and also makes the tests a bit harder to write and more of a change detector, so that's why I'm not enthusiastic about doing it that way.