Fix in place `ForDevice` function name; allow customizable FakeRoot name.
Since `ForDevice` takes in a device ID, just fix in place the function's
name. On the other hand, allow the root struct's name to be different,
and change its default to "root" as the name fits better with ypathgen's
purpose of path-building instead of ygen's purpose of representing device state.