Add option to use builder format for list path API
A threshold knob is added to the generator script to control from what
length on the builder API format should be used.

The idea is that having a huge combination of list path constructor
functions when the number of keys is large can be confusing to the
reader, and cause too much code to be generated: If there are 3 keys,
then there would be 2^3 = 8 different functions for the list, 20 would
mean 1 million.

Instead of forcing the user to select the right one among the masses,
the user can simply set the value of each key instead.