Add EnumeratedUnionSubsumingTypes in supplement to EnumeratedUnionTypes.
In #394, it has been agreed to name non-typedef enumerations residing in typedef unions as if they were typedefs themselves.

When figuring out whether an enumeration should be treated as a typedef, we need to know whether it's within a typedef union, and if it is, use the union typedef to name the enumeration.

Since goyang's `YangType` doesn't tells us whether or not it's a union subtype, we have to figure this out somehow.  I considered augmenting `YangType` to tell us in which type it resides, but I'm not sure whether `YangType`s can be reused as union subtypes, in which case there would be multiple such "subsuming" types.

In this PR, I define the "subsuming type" of each `YangType` to mean the closest/innermost type that's not a union subtype:
case 1: for any type not within a union, the subsuming type is itself.
case 2: for any type within a union, the subsuming type is the first containing union type that's a typedef, or highest union type definition, whichever one is closest to the type.

`EnumeratedUnionSubsumingTypes` extracts these "subsuming" types out of a union.