Allow gNMI paths with union keys embedded to be unmarshalled
Currently, union key values cannot be unmarshalled due to `StringToType`
not having a case statement handling the case when the string is to be
converted to a union type, causing an unmarshal error.

This change introduces handling for union types, which involves the
creation of a new function `StringToUnionType` due to the complexity in
handling unions.

The new handling function and helpers incorporates plenty of existing
logic, so some refactoring is done to reduce function size and reuse
code segments.