Transform leafref paths when prefer_operational_state is set.
Currently, the AST returned by goyang to ygot is never touched, including leafrefs that point to leaves under config.

This change modifies the AST's leafrefs to point to the state versions of the leaves when `prefer_operational_state` is set.

**Justification**:
When `prefer_operational_state` is set, ygot effectively prunes away the config portions of the YANG tree in the generated code representation. This means that whenever the schema is used for any processing (e.g. during code generation), the config path is used even though the node doesn't exist in the generated code.

One problem this causes is that the path tag in the generated Go code doesn't contain the leafref path (because it points to a config leaf, which cannot be found during processing due to being pruned). The result is config leafrefs cannot be unmarshalled.