Add flag ignore_shadowed_schema_paths (#429)
* Add flag keep_shadowed_schema_paths

This flag keeps the shadowed fields (i.e. state leaves for
PreferIntendedConfig, and config leaves for PreferOperationalState) as
part of the schema path within the generated code (e.g. `Path` tag for Go code).

This means config/state leaves that have been compressed out in the
compressed code can continue to unmarshal and marshal within the
generated code.

This flag is only relevant when compressPaths=true.

This flag works by adding an extra field in the IR `Directory` type
called `ShadowedFields`, mirroring `Fields`. It stores the
compressed-out (aka shadowed) fields and allow the extra paths to be
output as field tags.