Fix default value generation in leaf getters.
NOTE: This PR only fixes the generated code with the simple union flag
turned on. It doesn't address the wrapper union generated code.

The current exceedingly simple solution actually works for most cases,
but for unions, there are two challenges:
1. Each union subtype needs to be examined individually to see whether
it conforms, and the first one that does should be converted
accordingly. The current solution can't tell if `VALID_IMMEDIATELY`, for
example, represents a string or an enum, unless each one is tried.
Unfortunately, this means we need to do exhausive checking (since goyang
doesn't do this for us), similar to the checks we do when unmarshalling
unions.
2. Union values need to be wrapped within union types.

**Question**: Are leafrefs allowed to have default values? pyang doesn't
error out in this case. leafrefs are allowed within unions as of YANG
1.1 -- so we need to wrap the type if it's within a union somehow. I
chose to implement this using an auxiliary function -- I didn't like the
idea of intercepting the return value with a `defer`.

Converting to `Binary` is the other corner case addressed.