Fix default value generation in leaf getters. (#496)
* Fix default value generation in leaf getters.

NOTE: This PR only fixes the generated code with the simple union flag
turned on. It doesn't address the wrapper union generated code.

The current exceedingly simple solution actually works for most cases,
but for unions, there are two challenges:
1. Each union subtype needs to be examined individually to see whether
it conforms, and the first one that does should be converted
accordingly. The current solution can't tell if `VALID_IMMEDIATELY`, for
example, represents a string or an enum, unless each one is tried.
Unfortunately, this means we need to do exhausive checking (since goyang
doesn't do this for us), similar to the checks we do when unmarshalling
unions.
2. Union values need to be wrapped within union types.

Converting to `Binary` is the other corner case addressed.

Restriction validation functions are factored out of ytypes and used in
`yangDefaultValueToGo`, which now checks types in order, but does so by
checking not only a type match against the string, but also that the
value doesn't violate restrictions.