Don't initialize leaf values when unmarshalling valid shadow paths.
Currently when `SetNode` is called on a path, missing elements are initialized as the function traverses through the `GoStruct`. This happens even for paths leading to a leaf as well as shadow paths, causing the default value to always be populated into these elements.

This change prevents `SetNode` from initializing a leaf value when traversing a shadow path.