Always pair the leafref path with only the referred leaf in the struct tags for compressed schemas.
Currently for compressed schema GoStructs, we always put the key path in the `path` annotation of the struct tags, even if the leafref refers to the node under the config, and preferOperationalState=true, e.g. `path:"index|state/index" shadow-path:"config/index"`. However, if `index` points to `../config/index`, `index` doesn't really belong with `state/index`.

Instead, a more natural annotation would be `path:"state/index" shadow-path:"index|config/index"`.

This PR proposes to pair the leafref path with only the node that it points to.

This change leads to these behavioural changes:
- When unmarshalling datapoints to the annotation above, paths to the config-referring "index" are ignored. This makes sense since this isn't a state data anyways.
- When marshalling datapoints with the above annotation, and with reverseShadowPaths=true, the key gets marshalled, which is correct since it is part of the configuration. Conversely, when marshalling with reverseShadowPaths=false, the key path is omitted.