[Go]: Resolve conflict when a multi-keyed list contains a container named "key"
Do this by keeping the current naming, but change to "%s_%s_key" instead
of "%s_%s_Key" as the back up name if a conflict like this occurs. This
should always work since child non-leaves are named using camel cases,
so I'm making it error out if the backup is already taken.

Tested:
- Added a new integration test for this case.