SetNode: Don't Initialize Leaf Path if Failed when `InitMissingElements` is set.
A leaf path is currently unconditionally initialized for `SetNode` when `InitMissingElements` is set. If `SetNode` failed with this option set, then this not only doesn't make sense, but would also create an inconsistency against when this option is not set.

This PR keeps leaf initialization behaviour for `GetOrCreateNode`, but removes it for `SetNode` when `InitMissingElements` is set.