Use [reflect.DeepEqual](https://pkg.go.dev/reflect#DeepEqual) instead of cmp.Equal.
From [cmp](https://pkg.go.dev/github.com/google/go-cmp/cmp):
```
This package is intended to be a more powerful and safer alternative to
reflect.DeepEqual for comparing whether two values are semantically
equal. It is intended to only be used in tests, as performance is not a
goal and it may panic if it cannot compare the values. Its propensity
towards panicking means that its unsuitable for production environments
where a spurious panic may be fatal.
```