as commented by @wenovus in https://github.com/openconfig/ygot/issues/623
The fix is essentially correct: we handle the case of lists (a single level in YANG)
being represented as two levels in Go (map -> element), but don't handle the case of
leaf-lists also being represented as two levels in Go (slice -> element).
The "two levels" is how ForEachField processes elements, which creates this
corner case for DataNodeAtPath.

It turns out we already had a test case for this but it had an extra "../" s
o it was testing the wrong behaviour.