Do not trim compressed path elements for maps or slices.
Currently, we trim the path for a `map[]{}` or slice type in Go for
lists in order to prevent the list's name from appearing twice in the
data tree while traversing a list element. This is because
`forEachFieldInternal` currently creates new `NodeInfo` elements,
copying the list name, when traversing list elements.

Based on the tests that fail, this behaviour apparently tries to solve
the problem of processing "../" when calling `ytypes.dataNodesAtPath`,
but the logic in there incorrectly identifies compression to be the
issue: compression is indeed not the issue, but rather how slices and
maps are processed in two levels -- slices and maps exist equally in
compressed and uncompressed generated code.

This PR removes the dependency of existing logic on schema compression,
but rather on whether the current element is a map or a slice,
corresponding to a YANG list or leaf-list, and in such cases, it skips
that extra level created by `forEachFieldInternal` instead. This results
in slightly simpler, and much more understandable logic with the comment
updates.