Presence Containers (#645)
* This commit adds support for YANG presence containers. It adds a tag to
the field of generated Go Struct. When rendering JSON documents, before
deleting/omitting empty containers, it checks for this struct field tag.

As per the OpenConfig style guide, YANG presence containers are not
used, as such this would only be used with 3rd party YANG models. As
such a CLI knob is added to generator.go to enable this behavior with
-yangpresence=true

* fix broken test case because of resync with master

* Changes Requested #1 in https://github.com/openconfig/ygot/pull/645

* Adding https://github.com/openconfig/ygot/pull/645 requested changes:
- adding tests in ygot/render_test.go:
  - testing a presence container in a list entry (ISIS overload usecase)
  - testing a presence container in a non-presence container
    (system/ssh-server usecase)

- adding tests in ytypes/container_test.go:
  - adding tests in TestUnmarshalContainer

* Updated minor comments in https://github.com/openconfig/ygot/pull/645