Make `EnumeratedYANGType.IdentityBaseName` exported. (#688)
* Make `EnumeratedYANGType.IdentityBaseName` exported.

I think it makes sense for this to be exported. Not sure why I wanted it
to be extracted from the plugin originally.

* Add `LangMapperExt` and `Flags` fields to each level of the `IR`. (#689)

* Add `LangMapperExt` and `Flags` fields to each level of the `IR`.

The idea is to to call a different method from the interface for each level of hierarchy of the IR to populate the `Flag` fields for that level of the IR. So within `getOrderedDirDetails` for each `ParsedDirectory` and `NodeDetails`, and within the loop of genEnums (in `GenerateIR`) for each `EnumeratedYANGType`. This way the caller of `GenerateIR` is able to produce extra information for any level of the `IR`.

Tests added in genir_test.go

* Improve comments

* Generate `YANGType` using `LangMapperExt` mechanism. (#690)

* Generate `YANGType` using `LangMapperExt` mechanism.

* Remove dependency of enumgen on MappedType (#691)

* Remove dependency of enumgen on MappedType

* Change enumeratedTypedefTypeName to return a bool for whether the input type is actually a typedef

* Move enumset and schematree into `LangMapperBase`. (#692)

* Move enumset and schematree into `LangMapperBase`.

Currently, the individual LangMapper implementation instances (GoLangMapper and ProtoLangMapper) are required to hold these types, which can be confusing to the user since these are not fundamentally part of the LangMapper API -- rather it's their methods (i.e. enum name look-up and leafref resolution) that are part of the LangMapper API interface.

Embedding a special interface (`LangMapperBase`) not only decouples the implementation from the interface, but also provides a clear forward compatibility contract of methods (incl. enum name look-up and leafref resolution) available to implementors should more functionality become available in the future.

The proposed vision is:
* `LangMapper`-defined methods need to be implemented by LangMapper implementors.
* `LangMapperBaseSetup`-defined methods are set-up methods that are called either within `GenerateIR`, or by the unit tests of LangMapper methods.
* `LangMapperBase`-defined methods NOT part of the overall `LangMapper` interface are helper methods available to `LangMapper` implementors in their method implementations.

I will modify the design doc once all is agreed-upon and merged.

* Add `InjectEnumSet` and `InjectSchemaTree` to `LangMapperBase`. (#693)

* Add `SetupEnumSet` and `SetupSchemaTree` to `LangMapperBase`.

This allows unit tests to be created by custom LangMappers on their methods (e.g. `FieldName`, `KeyLeafType`).

* Rename Setup to Inject and fix-up comments

* Delete `YANGCodeGenerator.GetDirectoriesAndLeafTypes` (#694)

* Delete `YANGCodeGenerator.GetDirectoriesAndLeafTypes`

* Use bash instead of overalls (#698)

* Use bash instead of overalls

* Don't install overalls