`ytypes.DeleteNode` on a (single/multi) list key deletes the entire list entry.
Alternatively, we can add an option for this feature, and reject such
deletes instead.

Background:

Currently, `ytypes.DeleteNode` just deletes whatever path is
specified, so if it's a list key it deletes the key field. This creates
two issues:
1. Such a deletion makes the GoStruct invalid. It doesn't make sense
   from a schema perspective to only delete the key of a list.
2. If the user wanted to maintain an in-memory store of a list, then
   such a deletion would cause any future deletions on non-key fields of
   the same list entry to fail due to the key field not existing (this
   is an implementation detail, but essentially the current implementation
   assumes that the list is schema-valid, so the implementation is broken).

This PR proposes that `DeleteNode` deletes the list entry altogether
when this is issued.

It also brings a related question: What should a gNMI `SetRequest`'s
`delete` on key paths do? Should it have the same behaviour? Or should
it simply be rejected?