Add option to have `ytypes.DeleteNode` on a (single/multi) list key delete the entire list entry.
Currently, `ytypes.DeleteNode` just deletes whatever path is specified,
so if it's a list key it deletes the key field. This makes the GoStruct
invalid, which is useful for checking the validity of a `SetRequest`.

 Use Case
However, the reference gnmi cache implementation sends notifications for
all descendants instead of the top-level deleted list entry. This means
a client can receive a delete to a key, followed by a delete to a
non-key. This is problematic because the implementation of `DeleteNode`
assumes that the list is valid according to the schema, and so breaks
when this happens. This option supports this use case by having the
entire list entry be deleted. It makes the assumption that the reference
gnmi cache only sends such deletions when the list entry is going to be
deleted anyways. This is a reasonable assumption for a use case like
lemming, where schema-validity is maintained throughout its operation.

--------------
**I haven't decided on whether I like this or simply changing DeleteNode
to allow arbitrary deletions.**