Make ordered maps a pointer in the parent GoStruct.
Initially I made the choice to use a non-pointer so that the user would
not need to initialize this field before using it.

The advantage is all methods can hang off of the type rather than
needing methods to hang off the parent type in order to nil-check and
initialize.

However, this creates problems that a pointer field solves:
1. Avoid needing value-receivers in order to make the field satisfy
   certain interfaces (e.g. ygot.GoStruct). Mixing receiver types is
   confusing for the reader.
2. Helps avoid awkward reflection code that need to access the pointer
   receiver methods since `CanAddr()=false` for non-pointer objects:
   https://go.dev/play/p/5vYVVUS-GfD
3. Users of ygot are used to declaring every struct as pointers, so
   using pointers makes things congruent with existing generated code.