EmitJSON/ConstructIETFJSON/ConstructInternalJSON support for structs containing ordered maps.
Note: ordered maps themselves cannot be processed by these functions.

Ideally Go supports type constraints of unions of interfaces but it
doesn't, and I think it is too early to change the input parameters of
these functions from ygot.GoStruct -> interface{} or add even more
methods to achieve what ygnmi does with the query types -- which would
bloat the generated code for a small gain which might not be so useful
right now.