fix(goelements): handle Hex default int/uint values (#827)
* fix(goelements): handle Hex default int/uint values

* fix(goelements): handle Hex default int/uint values

* fix(gogen): fix hex/octal default value handling for int/uint

.. and add test cases for hex/octal default int/uint values.

The Go standard library already handles base inference for hexadecimal,
octal, binary, and base-10, with sign prefix support. There's no need to
reimplement the parsing and conversion of these integer values.

According to https://datatracker.ietf.org/doc/html/rfc7950#section-9.2.1
one remaining question is, should the `yangDefaultValueToGo` function
return an error if the default value for an integer is in binary format?

* fix(gogen): make sure that base 2 default int/uint values are disallowed

* fix(gogen): make sure that underscores are disallowed in default values

* fix(gogen): handle default int/uint negative case with a sign prefix

* fix(gogen): correct a typo on trimming the default value string

The tests didn't catch this typo because `ParseUint` returned an error
on the sign prefix. But this should be fixed and the expected error
should be `binary format is disallowed`.

* fix(gogen): fix typos in the default value parsing test cases