Move some of the test GoStructs to `testutil` so `util` package can use it
Ideally we generate this instead of handcrafting/copying it as it is
subject to change and we want to be able to catch that.

However the problem is because the generated code imports from ygot, it
creates a circular dependency. There might be a way to address it but it
should be made in a separate PR. This PR works around this by having a
handcrafted copy that doesn't import from ygot.

Later I will take a look at whether it's possible to move those ygot
definitions to a base package and doing an alias.